Move default margin for bar charts to different place
