+++ front/app/components/admin/Graphs/BarChart/index.tsx
@@ -8,9 +8,6 @@ import MultiBarChart, {
 // utils
 import { convertMapping, convertBarProps, Mapping, BarProps } from './utils';
 
-// typings
-import { Margin } from 'components/admin/Graphs/MultiBarChart/utils';
-
 interface Props extends Omit<MultiBarChartProps, 'mapping' | 'bars'> {
   mapping?: Mapping;
   bars?: BarProps;
@@ -30,10 +27,3 @@ const BarChart = ({ mapping, bars, ...otherProps }: Props) => {
 };
 
 export default BarChart;
-
-export const DEFAULT_MARGIN: Margin = {
-  top: 20,
-  right: 30,
-  left: 10,
-  bottom: 5,
-};
