+++ front/app/components/admin/Graphs/constants.ts
@@ -1,0 +1,13 @@
+export interface Margin {
+  top?: number;
+  bottom?: number;
+  left?: number;
+  right?: number;
+}
+
+export const DEFAULT_BAR_CHART_MARGIN: Margin = {
+  top: 20,
+  right: 30,
+  left: 10,
+  bottom: 5,
+};
