+++ front/app/containers/Admin/dashboard/users/charts/BarChartByCategory.tsx
@@ -16,7 +16,8 @@ import {
   GraphCard,
   GraphCardInner,
 } from 'components/admin/GraphWrappers';
-import BarChart, { DEFAULT_MARGIN } from 'components/admin/Graphs/BarChart';
+import BarChart from 'components/admin/Graphs/BarChart';
+import { DEFAULT_BAR_CHART_MARGIN } from 'components/admin/Graphs/constants';
 import { Tooltip, LabelList } from 'recharts';
 
 // resources
@@ -98,7 +99,7 @@ export class BarChartByCategory extends React.PureComponent<
           <BarChart
             data={serie}
             innerRef={this.currentChart}
-            margin={DEFAULT_MARGIN}
+            margin={DEFAULT_BAR_CHART_MARGIN}
             bars={{ name: unitName }}
             renderLabels={(props) => <LabelList {...props} position="top" />}
             renderTooltip={(props) => <Tooltip {...props} />}
